backend/requirements.txt
fastapi>=0.135
uvicorn>=0.42
python-multipart>=0.0.22
openai>=2.29
python-dotenv>=1.2
